Reorganize conn/credentials
